build: trigger documentation generation within CMake
This means we don't need some other doc-generation top-level script.

Adds two new flags
DOC_AUDIENCE determines what is included in the doc bundle:
- APP builds the docs for embedding in the help system
- WEB builds the docs for deployment to GH pages, and includes WASM targets
- DEV builds the docs, including API and design docs.

DOC_PREBUILT_DIR allows documentation generation to be skipped.
If set, this variable is assumed to point to a directory which contains valid documentation.

When building WEB docs, you probably want to be in a WASM environment.
Otherwise, no demos will be generated.

Implements all features needed by #365 and #366, other than CI.

+++ docs/copy_dirs.py
@@ -26,6 +26,7 @@ def setup(app):
 def helper(exts, absSrcDir, absDst):
   # logger.info(f"Copying {absSrcDir} to {absDst}")
   # logger.info(f'Copying {file} to {dir}')#insert into innermost loop.
+  if not absDst.exists(): return
   for ext in exts:
     for file in pathlib.Path(absSrcDir).glob(f"{ext}"): shutil.copy(file, absDst)
 
@@ -40,4 +41,4 @@ def copy_dirs(app, exception):
       pair = pair.split("?")
       pairs.append([pathlib.Path(app.srcdir)/pair[0], pathlib.Path(app.outdir)/pair[1]])
 
-    for pair in pairs: helper(app, *pair)
+    for pair in pairs: helper(app.config.exts, *pair)
